feat(manage): add notification logic for app received email
