+++ apps/manage/src/app/views/cases/view/controller.test.js
@@ -1,10 +1,8 @@
 import { describe, it, mock } from 'node:test';
 import assert from 'node:assert';
-import { buildGetJourneyMiddleware, buildUpdateCase, buildViewCaseDetails } from './controller.js';
+import { buildGetJourneyMiddleware, buildViewCaseDetails } from './controller.js';
 import { configureNunjucks } from '../../../nunjucks.js';
 import { mockLogger } from '@pins/crowndev-lib/testing/mock-logger.js';
-import { APPLICATION_PROCEDURE_ID } from '@pins/crowndev-database/src/seed/data-static.js';
-import { Prisma } from '@prisma/client';
 import { assertRenders404Page } from '@pins/crowndev-lib/testing/custom-asserts.js';
 
 describe('case details', () => {
@@ -348,304 +346,6 @@ describe('case details', () => {
 			const viewData = mockRes.render.mock.calls[0].arguments[1];
 			assert.ok(viewData);
 			assert.strictEqual(viewData.sharePointLink, 'https://example.com');
-		});
-	});
-	describe('buildUpdateCase', () => {
-		it('should throw if no id', async () => {
-			const updateCase = buildUpdateCase({});
-			const mockReq = { params: {} };
-			const mockRes = { locals: {} };
-			const data = {};
-			await assert.rejects(() => updateCase({ req: mockReq, res: mockRes, data }));
-		});
-		it('should do nothing if no updates', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					update: mock.fn()
-				}
-			};
-			const updateCase = buildUpdateCase({ db: mockDb, logger });
-			const mockReq = { params: { id: 'case-1' } };
-			const mockRes = { locals: {} };
-			const data = {};
-			await updateCase({ req: mockReq, res: mockRes, data });
-			assert.strictEqual(mockDb.crownDevelopment.update.mock.callCount(), 0);
-			assert.strictEqual(logger.info.mock.callCount(), 2);
-			const args = logger.info.mock.calls[1].arguments[1];
-			assert.strictEqual(args, 'no case updates to apply');
-		});
-		it('should call db update and add to session', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					update: mock.fn()
-				}
-			};
-			const updateCase = buildUpdateCase({ db: mockDb, logger });
-			const mockReq = {
-				params: { id: 'case1' },
-				session: {}
-			};
-			const mockRes = { locals: {} };
-			/** @type {{answers: import('./types.js').CrownDevelopmentViewModel}} */
-			const data = {
-				answers: {
-					description: 'My new application description'
-				}
-			};
-			await updateCase({ req: mockReq, res: mockRes, data });
-			assert.strictEqual(mockDb.crownDevelopment.update.mock.callCount(), 1);
-			const updateArg = mockDb.crownDevelopment.update.mock.calls[0].arguments[0];
-			assert.strictEqual(updateArg.where?.id, 'case1');
-			assert.strictEqual(mockReq.session?.cases?.case1.updated, true);
-		});
-		it('should fetch case data from the journey for relation IDs', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					update: mock.fn()
-				}
-			};
-			const updateCase = buildUpdateCase({ db: mockDb, logger });
-			const mockReq = {
-				params: { id: 'case1' },
-				session: {}
-			};
-			const mockRes = {
-				locals: {
-					journeyResponse: {
-						answers: {
-							eventId: 'event-1',
-							procedureId: APPLICATION_PROCEDURE_ID.INQUIRY
-						}
-					}
-				}
-			};
-			/** @type {{answers: import('./types.js').CrownDevelopmentViewModel}} */
-			const data = {
-				answers: {
-					inquiryVenue: 'some place'
-				}
-			};
-			await updateCase({ req: mockReq, res: mockRes, data });
-			assert.strictEqual(mockDb.crownDevelopment.update.mock.callCount(), 1);
-			assert.strictEqual(mockReq.session?.cases?.case1.updated, true);
-			const updateArg = mockDb.crownDevelopment.update.mock.calls[0].arguments[0];
-			assert.strictEqual(updateArg.where?.id, 'case1');
-			assert.strictEqual(updateArg.data?.Event?.upsert?.where?.id, 'event-1');
-			assert.strictEqual(updateArg.data?.Event?.upsert?.create.venue, 'some place');
-		});
-		it('should dispatch Lpa Acknowledge Receipt Of Questionnaire Notification with site address', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					findUnique: mock.fn(() => ({
-						id: 'case-1',
-						reference: 'CROWN/2025/0000001',
-						description: 'a big project',
-						siteAddressId: 'address-1',
-						SiteAddress: { line1: '4 the street', line2: 'town', postcode: 'wc1w 1bw' },
-						Lpa: { email: 'test@email.com' }
-					})),
-					update: mock.fn()
-				}
-			};
-			const mockNotifyClient = {
-				sendLpaAcknowledgeReceiptOfQuestionnaire: mock.fn()
-			};
-			const updateCase = buildUpdateCase({
-				db: mockDb,
-				logger,
-				notifyClient: mockNotifyClient,
-				portalBaseUrl: 'https://test.com'
-			});
-			const mockReq = {
-				params: { id: 'case1' },
-				session: {}
-			};
-			const mockRes = {
-				locals: {
-					journeyResponse: {
-						answers: {
-							lpaQuestionnaireReceivedEmailSent: false
-						}
-					}
-				}
-			};
-			const date = new Date('2025-01-02');
-			/** @type {{answers: import('./types.js').CrownDevelopmentViewModel}} */
-			const data = {
-				answers: {
-					lpaQuestionnaireReceivedDate: date
-				}
-			};
-
-			await updateCase({ req: mockReq, res: mockRes, data });
-			assert.strictEqual(mockDb.crownDevelopment.update.mock.callCount(), 1);
-			assert.strictEqual(mockReq.session?.cases?.case1.updated, true);
-
-			const updateArg = mockDb.crownDevelopment.update.mock.calls[0].arguments[0];
-			assert.strictEqual(updateArg.where?.id, 'case1');
-			assert.strictEqual(updateArg.data?.lpaQuestionnaireReceivedDate, date);
-			assert.strictEqual(updateArg.data?.lpaQuestionnaireReceivedEmailSent, true);
-
-			assert.strictEqual(mockNotifyClient.sendLpaAcknowledgeReceiptOfQuestionnaire.mock.callCount(), 1);
-			assert.deepStrictEqual(mockNotifyClient.sendLpaAcknowledgeReceiptOfQuestionnaire.mock.calls[0].arguments, [
-				'test@email.com',
-				{
-					reference: 'CROWN/2025/0000001',
-					applicationDescription: 'a big project',
-					siteAddress: '4 the street, town, wc1w 1bw',
-					lpaQuestionnaireReceivedDate: '2 Jan 2025',
-					frontOfficeLink: 'https://test.com/applications'
-				}
-			]);
-		});
-		it('should dispatch Lpa Acknowledge Receipt Of Questionnaire Notification with northing/easting', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					findUnique: mock.fn(() => ({
-						id: 'case-1',
-						reference: 'CROWN/2025/0000001',
-						description: 'a big project',
-						siteNorthing: '123456',
-						siteEasting: '654321',
-						Lpa: { email: 'test@email.com' }
-					})),
-					update: mock.fn()
-				}
-			};
-			const mockNotifyClient = {
-				sendLpaAcknowledgeReceiptOfQuestionnaire: mock.fn()
-			};
-			const updateCase = buildUpdateCase({
-				db: mockDb,
-				logger,
-				notifyClient: mockNotifyClient,
-				portalBaseUrl: 'https://test.com'
-			});
-			const mockReq = {
-				params: { id: 'case1' },
-				session: {}
-			};
-			const mockRes = {
-				locals: {
-					journeyResponse: {
-						answers: {
-							lpaQuestionnaireReceivedEmailSent: false
-						}
-					}
-				}
-			};
-			const date = new Date('2025-01-02');
-			/** @type {{answers: import('./types.js').CrownDevelopmentViewModel}} */
-			const data = {
-				answers: {
-					lpaQuestionnaireReceivedDate: date
-				}
-			};
-
-			await updateCase({ req: mockReq, res: mockRes, data });
-			assert.strictEqual(mockDb.crownDevelopment.update.mock.callCount(), 1);
-			assert.strictEqual(mockReq.session?.cases?.case1.updated, true);
-
-			const updateArg = mockDb.crownDevelopment.update.mock.calls[0].arguments[0];
-			assert.strictEqual(updateArg.where?.id, 'case1');
-			assert.strictEqual(updateArg.data?.lpaQuestionnaireReceivedDate, date);
-			assert.strictEqual(updateArg.data?.lpaQuestionnaireReceivedEmailSent, true);
-
-			assert.strictEqual(mockNotifyClient.sendLpaAcknowledgeReceiptOfQuestionnaire.mock.callCount(), 1);
-			assert.deepStrictEqual(mockNotifyClient.sendLpaAcknowledgeReceiptOfQuestionnaire.mock.calls[0].arguments, [
-				'test@email.com',
-				{
-					reference: 'CROWN/2025/0000001',
-					applicationDescription: 'a big project',
-					siteAddress: 'Northing: 123456 , Easting: 654321',
-					lpaQuestionnaireReceivedDate: '2 Jan 2025',
-					frontOfficeLink: 'https://test.com/applications'
-				}
-			]);
-		});
-		it('should throw error if notification dispatch fails', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					findUnique: mock.fn(() => ({
-						id: 'case-1',
-						reference: 'CROWN/2025/0000001',
-						description: 'a big project',
-						siteNorthing: '123456',
-						siteEasting: '654321',
-						Lpa: { email: 'test@email.com' }
-					})),
-					update: mock.fn()
-				}
-			};
-			const mockNotifyClient = {
-				sendLpaAcknowledgeReceiptOfQuestionnaire: mock.fn(() => {
-					throw new Error('Exception error');
-				})
-			};
-			const updateCase = buildUpdateCase({
-				db: mockDb,
-				logger,
-				notifyClient: mockNotifyClient,
-				portalBaseUrl: 'https://test.com'
-			});
-			const mockReq = {
-				params: { id: 'case1' },
-				session: {}
-			};
-			const mockRes = {
-				locals: {
-					journeyResponse: {
-						answers: {
-							lpaQuestionnaireReceivedEmailSent: false
-						}
-					}
-				}
-			};
-			const date = new Date('2025-01-02');
-			/** @type {{answers: import('./types.js').CrownDevelopmentViewModel}} */
-			const data = {
-				answers: {
-					lpaQuestionnaireReceivedDate: date
-				}
-			};
-
-			await assert.rejects(() => updateCase({ req: mockReq, res: mockRes, data }));
-		});
-		it('should not throw Prisma errors', async () => {
-			const logger = mockLogger();
-			const mockDb = {
-				crownDevelopment: {
-					update: mock.fn(() => {
-						throw new Prisma.PrismaClientKnownRequestError('Error', { code: 'E101' });
-					})
-				}
-			};
-			const updateCase = buildUpdateCase({ db: mockDb, logger });
-			const mockReq = {
-				params: { id: 'case1' },
-				session: {}
-			};
-			const mockRes = { locals: {} };
-			/** @type {{answers: import('./types.js').CrownDevelopmentViewModel}} */
-			const data = {
-				answers: {
-					description: 'My new application description'
-				}
-			};
-			await assert.rejects(
-				() => updateCase({ req: mockReq, res: mockRes, data }),
-				(err) => {
-					assert.strictEqual(err.name, 'Error');
-					assert.strictEqual(err.message, 'Error updating case (E101)');
-					return true;
-				}
-			);
 		});
 	});
 });
