+++ infrastructure/variables.tf
@@ -49,10 +49,12 @@ variable "apps_config" {
     gov_notify = object({
       disabled = bool
       templates = object({
-        test_template_id    = string
-        pre_ack_template_id = string
-        ack_rep_template_id = string
-        lpa_qnr_template_id = string
+        test_template_id                = string
+        pre_ack_template_id             = string
+        ack_rep_template_id             = string
+        lpa_qnr_template_id             = string
+        app_rec_with_fee_template_id    = string
+        app_rec_without_fee_template_id = string
       })
     })
